Yoast Banner Remover
This plugin removes the sidebar ads from <a
href="https://wordpress.org/plugins/google-analytics-for-wordpress/">Goo
gle Analytics by Yoast</a> plugin and <a
href="https://wordpress.org/plugins/wordpress-seo/">Yoast SEO</a>
plugin. By no means this is not intended to be a lack of appreciation
for the plugin authors and we highly recommend using those plugins.
It's just that sometimes a distraction free dashboard is more suitable,
especially when dealing with clients.